requirements/prd.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
pydantic==2.11.7
face_recognition==1.3.0
numpy==2.3.0
scikit-learn==1.7.0
Pillow==11.2.1